dsp_backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
